Created new branch at commit 7372b75241d65dd23bde0e3372db03a8c7ae66fa This rolls the mock catalog generation and sampling back to 7/30/21 version. I have then moved over several modernization lines of code from the master branch version to update key functionality.
The major change between the branches is that the mock catalog generation pulls the completeness correction, agn selection membership, and the J-band absolute magnitude from the real data again while currently turning off any lines involving the LF in the sampler. These scripts should be able to recreate the mock catalog from 8/2/21 and recover a good fit to the data. The one exception to the 8/2/21 catalog is that this now has corrected J-band absolute magnitudes in the real data catalog which were wrong in the original version.